+++ src/commands/blob/blobContainerActionHandlers.ts
@@ -7,11 +7,12 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { IActionContext, registerCommand } from 'vscode-azureextensionui';
 import { AzureStorageFS } from '../../AzureStorageFS';
+import { ext } from '../../extensionVariables';
 import { storageExplorerLauncher } from '../../storageExplorerLauncher/storageExplorerLauncher';
 import { BlobContainerTreeItem } from '../../tree/blob/BlobContainerTreeItem';
 import { BlobDirectoryTreeItem } from '../../tree/blob/BlobDirectoryTreeItem';
 import { BlobTreeItem } from '../../tree/blob/BlobTreeItem';
-import { createBlobDirectory, findBlobTreeItem, getBlobResourceType, IBlobContainerCreateChildContext, showBlobPathInputBox } from '../../utils/blobUtils';
+import { IBlobContainerCreateChildContext, showBlobPathInputBox } from '../../utils/blobUtils';
 import { localize } from '../../utils/localize';
 import { deleteNode } from '../commonTreeCommands';
 
@@ -22,35 +23,32 @@ export function registerBlobContainerActionHandlers(): void {
     registerCommand("azureStorage.createBlockBlob", async (context: IActionContext, parent: BlobContainerTreeItem) => {
         const blobPath: string = await showBlobPathInputBox(parent);
         const dirNames: string[] = blobPath.includes('/') ? path.dirname(blobPath).split('/') : [];
-        let dirParentUri: vscode.Uri = AzureStorageFS.idToUri(parent.fullId);
+        let currentDirId: string = parent.fullId;
         let currentDirPath: string = '';
 
         for (const dirName of dirNames) {
             currentDirPath += dirName;
+            currentDirId = `${parent.fullId}/${currentDirPath}`;
 
-            try {
-                let fileType: vscode.FileType = await getBlobResourceType(currentDirPath, parent, context);
+            let treeItem: BlobTreeItem | BlobDirectoryTreeItem | BlobContainerTreeItem | undefined = await ext.tree.findTreeItem(currentDirId, context);
+            if (!treeItem) {
+                // This directory doesn't exist yet
+                const dirParentPath: string = currentDirPath.includes('/') ? `/${path.dirname(currentDirPath)}` : '';
+                const dirParentId: string = `${parent.fullId}${dirParentPath}`;
+                const dirParentTreeItem: BlobDirectoryTreeItem | BlobContainerTreeItem = <BlobDirectoryTreeItem | BlobContainerTreeItem>await ext.tree.findTreeItem(dirParentId, context);
+                await dirParentTreeItem.createChild(<IBlobContainerCreateChildContext>{ ...context, childType: BlobDirectoryTreeItem.contextValue, childName: dirName });
+            } else {
+                const fileType: vscode.FileType = treeItem instanceof BlobDirectoryTreeItem || treeItem instanceof BlobContainerTreeItem ? vscode.FileType.Directory : vscode.FileType.File;
                 if (fileType !== vscode.FileType.Directory) {
                     throw new Error(localize('resourceIsNotADirectory', `"${currentDirPath}" is not a directory`));
                 }
-            } catch (err) {
-                if (err instanceof vscode.FileSystemError) {
-                    // This directory doesn't exist yet
-                    const dirParentPath: string = currentDirPath.includes('/') ? path.dirname(currentDirPath) : '';
-                    const dirParentTreeItem: BlobDirectoryTreeItem | BlobContainerTreeItem = <BlobDirectoryTreeItem | BlobContainerTreeItem>await findBlobTreeItem(dirParentUri, dirParentPath, parent, context);
-                    await createBlobDirectory(currentDirPath, dirParentPath, dirParentTreeItem, context);
-                } else {
-                    throw err;
-                }
             }
 
-            dirParentUri = AzureStorageFS.idToUri(parent.fullId, currentDirPath);
             currentDirPath += '/';
         }
 
-        currentDirPath = currentDirPath ? currentDirPath.slice(0, -1) : ''; // Remove trailing slash if it exists
-        const blobParent: BlobDirectoryTreeItem | BlobContainerTreeItem = <BlobDirectoryTreeItem | BlobContainerTreeItem>await findBlobTreeItem(dirParentUri, currentDirPath, parent, context);
-        const childTreeItem: BlobTreeItem = <BlobTreeItem>await blobParent.createChild(<IBlobContainerCreateChildContext>{ ...context, childType: 'azureBlob', childName: blobPath });
+        const blobParent: BlobDirectoryTreeItem | BlobContainerTreeItem = <BlobDirectoryTreeItem | BlobContainerTreeItem>await ext.tree.findTreeItem(currentDirId, context);
+        const childTreeItem: BlobTreeItem = <BlobTreeItem>await blobParent.createChild(<IBlobContainerCreateChildContext>{ ...context, childType: BlobTreeItem.contextValue, childName: blobPath });
         await vscode.commands.executeCommand("azureStorage.editBlob", childTreeItem);
     });
     registerCommand("azureStorage.uploadBlockBlob", async (context: IActionContext, treeItem: BlobContainerTreeItem) => await treeItem.uploadBlockBlob(context));
