+++ src/utils/blobUtils.ts
@@ -8,7 +8,6 @@ import * as mime from "mime";
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { AzExtTreeItem, IActionContext, ICreateChildImplContext } from 'vscode-azureextensionui';
-import { AzureStorageBlobTreeItem, getFileSystemError } from '../AzureStorageFS';
 import { maxPageSize } from '../constants';
 import { ext } from '../extensionVariables';
 import { BlobContainerTreeItem } from '../tree/blob/BlobContainerTreeItem';
@@ -122,61 +121,6 @@ export async function showBlobPathInputBox(parent: BlobContainerTreeItem | BlobD
             return undefined;
         }
     });
-}
-
-export async function findBlobTreeItem(uri: vscode.Uri, filePath: string, containerTreeItem: BlobContainerTreeItem, context: IActionContext, endSearchEarly?: boolean): Promise<AzureStorageBlobTreeItem> {
-    let treeItem: BlobTreeItem | BlobDirectoryTreeItem | BlobContainerTreeItem = containerTreeItem;
-
-    if (filePath === '') {
-        return treeItem;
-    }
-
-    let pathToLook = filePath.split('/');
-    for (const childName of pathToLook) {
-        if (treeItem instanceof BlobTreeItem) {
-            if (endSearchEarly) {
-                return treeItem;
-            }
-            throw getFileSystemError(uri, context, vscode.FileSystemError.FileNotFound);
-        }
-
-        let children: AzExtTreeItem[] = await treeItem.getCachedChildren(context);
-        let child = children.find((element) => {
-            if (element instanceof BlobTreeItem) {
-                return element.blobName === childName;
-            } else if (element instanceof BlobDirectoryTreeItem) {
-                return element.dirName === childName;
-            }
-            return false;
-        });
-        if (!child) {
-            if (endSearchEarly) {
-                return treeItem;
-            }
-            throw getFileSystemError(uri, context, vscode.FileSystemError.FileNotFound);
-        }
-
-        treeItem = <BlobTreeItem | BlobDirectoryTreeItem>child;
-    }
-
-    return treeItem;
-}
-
-export async function getBlobResourceType(filePath: string, containerTreeItem: BlobContainerTreeItem, context: IActionContext): Promise<vscode.FileType> {
-    let treeItem: AzureStorageBlobTreeItem = await findBlobTreeItem(vscode.Uri.parse(containerTreeItem.fullId), filePath, containerTreeItem, context);
-    return treeItem instanceof BlobDirectoryTreeItem || treeItem instanceof BlobContainerTreeItem ? vscode.FileType.Directory : vscode.FileType.File;
-}
-
-export async function createBlobDirectory(fullDirPath: string, parentPath: string, parentTreeItem: BlobDirectoryTreeItem | BlobContainerTreeItem, context: IActionContext): Promise<void> {
-    let matches = fullDirPath.match(`^${regexEscape(parentPath)}\/?([^\/^]+)\/?(.*?)$`);
-    while (!!matches) {
-        parentTreeItem = <BlobDirectoryTreeItem>await parentTreeItem.createChild(<IBlobContainerCreateChildContext>{ ...context, childType: 'azureBlobDirectory', childName: matches[1] });
-        matches = matches[2].match("^([^\/]+)\/?(.*?)$");
-    }
-}
-
-function regexEscape(s: string): string {
-    return s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
 }
 
 export interface IBlobContainerCreateChildContext extends IActionContext {
